Rename ActiveModel::Model to ActiveModel::BasicModel
This module provides the sets of features required to support the ActiveModel API and no more. Its naming is meant to be reminiscent of Ruby's Object and BasicObject.
