+++ activemodel/README.rdoc
@@ -14,23 +14,21 @@ Model solves this by defining an explicit API. You can read more about the
 API in ActiveModel::Lint::Tests.
 
 Active Model provides a default module that implements the basic API required
-to integrate with Action Pack out of the box: <tt>ActiveModel::Model</tt>.
+to integrate with Action Pack out of the box: <tt>ActiveModel::BasicModel</tt>.
 
     class Person
-      include ActiveModel::Model
+      include ActiveModel::BasicModel
 
       attr_accessor :name, :age
       validates_presence_of :name
     end
 
-    person = Person.new(:name => 'bob', :age => '18')
-    person.name # => 'bob'
-    person.age # => '18'
-    person.valid? # => true
+    person = Person.new
+    person.valid? # => false
 
 It includes model name introspections, conversions, translations and
 validations, resulting in a class suitable to be used with Action Pack.
-See <tt>ActiveModel::Model</tt> for more examples.
+See <tt>ActiveModel::BasicModel</tt> for more examples.
 
 Active Model also provides the following functionality to have ORM-like
 behavior out of the box:
