+++ activemodel/lib/active_model.rb
@@ -28,6 +28,7 @@ module ActiveModel
   extend ActiveSupport::Autoload
 
   autoload :AttributeMethods
+  autoload :BasicModel
   autoload :BlockValidator, 'active_model/validator'
   autoload :Callbacks
   autoload :Configuration
@@ -37,7 +38,6 @@ module ActiveModel
   autoload :Errors
   autoload :Lint
   autoload :MassAssignmentSecurity
-  autoload :Model
   autoload :Name, 'active_model/naming'
   autoload :Naming
   autoload :Observer, 'active_model/observing'
